requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-bootstrap-v5==1.0.4
django-bootstrap5==2.0.1
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.1
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
whitenoise==5.2.0
zipp==3.4.1
